+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * new Sepia avatars
   * new Cats avatars
   * new Birds avatars
+  * new Fenecs avatars
   * plugin settings to change the avatar set (for now it is only possible to choose on the instance level, channel option will come later)
 
 ### Minor changes and fixes